feat: floating combat text, fps-independent dialog, lowercase server paths
- Add N| floating damage/miss text for all combat scenarios (PvP melee,
  PvP ranged, NPC→User, User→NPC miss, spell damage on users)
- Fix dialog text animation to be frame-rate independent using delta-time
  scaling (was tied to FPS, too fast at high frame rates)
- Rename server data directories and config files to lowercase:
  Dat→dat, Dioses→dioses, Maps→maps, Server.ini→server.ini,
  HD.ini→hd.ini, Facciones.ini→facciones.ini, Configuracion.ini→configuracion.ini
- Update all Rust source references to match new lowercase paths
- Add "Borrar Personaje" button to character selection screen with
  confirmation dialog (random 4-digit code) sending TBRP packet

+++ client/Scripts/Game/Character.cs
@@ -52,9 +52,9 @@ public class Character
     public string DialogText = "";
     public string DialogColor = "FFFFFF";
     public long DialogStartMs;       // Environment.TickCount64 when created
-    public long DialogDurationMs;    // 5000 + 100 * text.Length
-    public int DialogRiseCounter;    // VB6 Sube: 18→0, decrements each tick
-    public int DialogAlpha;          // VB6 Desvanecimiento: starts 20, +12/frame while Sube>0, -10/frame on fade
+    public long DialogDurationMs;    // 3000 + 50 * text.Length
+    public float DialogRiseCounter;  // VB6 Sube: 18→0, delta-based (60/sec at VB6 rate)
+    public float DialogAlpha;        // VB6 Desvanecimiento: starts 20, +720/sec while rising, -600/sec on fade
     public bool DialogFading;        // VB6 Tiempito: True when lifetime expired, fading out
 
     // Auras (VB6: 5 equipment slots + 1 NPC aura)
